requirements.txt
asgiref==3.3.4
Django==3.2
django-cors-headers==3.7.0
djangorestframework==3.12.4
mysqlclient==2.0.3
paytmchecksum==1.7.0
pkg-resources==0.0.0
pycryptodome==3.10.1
PyMySQL==1.0.2
pytz==2021.1
sqlparse==0.4.1
typing-extensions==3.7.4.3
